Techtrees

See merge request cmdevs/colonial-warfare!996

+++ code/modules/cm_aliens/weeds.dm
@@ -1,6 +1,7 @@
 #define WEED_BASE_GROW_SPEED (10 SECONDS)
 #define WEED_BASE_DECAY_SPEED (20 SECONDS)
 
+
 /obj/effect/alien/weeds
 	name = "weeds"
 	desc = "Weird black weeds..."
@@ -19,6 +20,7 @@
 
 	var/datum/hive_status/linked_hive = null
 	var/hivenumber = XENO_HIVE_NORMAL
+	var/turf/weeded_turf
 
 	// Which node is responsible for keeping this weed patch alive?
 	var/obj/effect/alien/weeds/node/parent = null
@@ -44,6 +46,22 @@
 	update_neighbours()
 	if(node && node.loc && (get_dist(node, src) < node.node_range) && !hibernate)
 		addtimer(CALLBACK(src, .proc/weed_expand), WEED_BASE_GROW_SPEED / max(weed_strength, 1))
+
+	var/turf/T = get_turf(src)
+	T.weeds = src
+	weeded_turf = T
+
+	RegisterSignal(src, list(
+		COMSIG_ATOM_TURF_CHANGE,
+		COMSIG_MOVABLE_TURF_ENTERED
+	), .proc/set_turf_weeded)
+
+/obj/effect/alien/weeds/proc/set_turf_weeded(var/datum/source, var/turf/T)
+	SIGNAL_HANDLER
+	if(weeded_turf)
+		weeded_turf.weeds = null
+
+	T.weeds = src
 
 /obj/effect/alien/weeds/initialize_pass_flags(var/datum/pass_flags_container/PF)
 	. = ..()
@@ -93,6 +111,11 @@
 
 	var/oldloc = loc
 	parent = null
+
+	if(weeded_turf)
+		weeded_turf.weeds = null
+
+	weeded_turf = null
 	. = ..()
 	update_neighbours(oldloc)
 
@@ -145,7 +168,9 @@
 
 		var/obj/effect/alien/weeds/W = locate() in T
 		if(W)
-			if(W.weed_strength >= WEED_LEVEL_HIVE)
+			if(W.indestructible)
+				continue
+			else if(W.weed_strength >= WEED_LEVEL_HIVE)
 				continue
 			else if (W.linked_hive == node.linked_hive && W.weed_strength >= node.weed_strength)
 				continue
@@ -248,6 +273,9 @@
 		overlays += secretion
 
 /obj/effect/alien/weeds/ex_act(severity)
+	if(indestructible)
+		return
+
 	switch(severity)
 		if(0 to EXPLOSION_THRESHOLD_LOW)
 			if(prob(50))
@@ -259,6 +287,9 @@
 			qdel(src)
 
 /obj/effect/alien/weeds/attack_alien(mob/living/carbon/Xenomorph/X)
+	if(indestructible)
+		return
+
 	if(!HIVE_ALLIED_TO_HIVE(X.hivenumber, hivenumber))
 		X.animation_attack_on(src)
 
@@ -271,6 +302,9 @@
 
 
 /obj/effect/alien/weeds/attackby(obj/item/W, mob/living/user)
+	if(indestructible)
+		return FALSE
+
 	if(QDELETED(W) || QDELETED(user) || (W.flags_item & NOBLUDGEON))
 		return 0
 
@@ -298,15 +332,24 @@
 	return TRUE //don't call afterattack
 
 /obj/effect/alien/weeds/proc/healthcheck()
+	if(indestructible)
+		return
+
 	if(health <= 0)
 		qdel(src)
 
 /obj/effect/alien/weeds/flamer_fire_act(dam)
+	if(indestructible)
+		return
+
 	. = ..()
 	if(!QDELETED(src))
 		QDEL_IN(src, rand(1 SECONDS, 2 SECONDS)) // 1-2 seconds
 
 /obj/effect/alien/weeds/acid_spray_act()
+	if(indestructible)
+		return
+
 	. = ..()
 	health -= 20 * WEED_XENO_DAMAGEMULT
 	healthcheck()
